+++ frontend/src/utils/encodingUtils.js
